renamed posts to post, added fb module and all app.
