+++ facebook/all/models.py
@@ -1,0 +1,7 @@
+
+from facebook.modules.profile.user.models import User, TestUser
+from facebook.modules.profile.page.models import Page
+from facebook.modules.profile.event.models import Event
+from facebook.modules.profile.application.models import Request, Score
+from facebook.modules.media.models import Photo
+from facebook.modules.connections.post.models import Post
