+++ facebook/fb.py
@@ -1,0 +1,6 @@
+""" Use this module for imports. from facebook import fb """
+
+from graph import GraphAPIError, get_graph, get_static_graph, get_public_graph
+from session import get_session
+from modules.profile.application.utils import get_app_dict
+from fql import get_FQL
