+++ facebook/modules/profile/application/models.py
@@ -5,7 +5,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from facebook.fields import JSONField
 from facebook.models import Base
-from facebook.profile.user.models import User
+from facebook.modules.profile.user.models import User
 from facebook.graph import get_graph, get_static_graph, GraphAPIError
